+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - AutoUpdater should now recognize pre-releases!
 - Modals now open exlusively, i.e you cannot open the Changelog and the License at the same time
+- List of open character sheets now saved when a new character is added or a tab is closed
 
 ## [v0.5.0] - 2019-11-25 
 ### Added